@@ -1,7 +1,7 @@
 import React, {Component} from "react";
 import PropTypes from 'prop-types';
-import Typography from '@material-ui/core/Typography';
-import {withStyles} from '@material-ui/core/styles';
+import Typography from '@mui/material/Typography';
+import {withStyles} from '@mui/material/styles';
 
 const styles = theme => ({
   container: {
